+++ src/test/java/co/wanted/board/api/auth/encode/application/PasswordEncoderTest.java
@@ -1,4 +1,4 @@
-package co.wanted.board.api.auth.applicaion;
+package co.wanted.board.api.auth.encode.application;
 
 import co.wanted.board.api.member.domain.Password;
 import lombok.extern.slf4j.Slf4j;
@@ -6,9 +6,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-import java.util.Base64;
 
 import static org.assertj.core.api.Assertions.assertThat;
