+++ src/test/java/co/wanted/board/api/member/application/SignupServiceTest.java
@@ -45,7 +45,5 @@ class SignupServiceTest {
 
             assertThat(memberRepository.existsByEmail(mail)).isTrue();
         }
-
-
     }
 }
